네비게이션 바 수정
