+++ JM_ERP/src/main/java/com/example/demo/Controller/SD_SaleController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.example.demo.Service.SD_PurchaseService;
+import com.example.demo.Service.SD_SaleService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,13 +18,14 @@ public class SD_SaleController {
 	
 	@Autowired
 	private final SD_PurchaseService purchaseService;
+	private final SD_SaleService saleService;
 
 
 // 인기상품 그래프 출력할 화면   
 	@GetMapping("/best")
 	public String bestGraph(Model model) {
-		String graphHtmlPath = purchaseService.executePythonScript();
+		String graphHtmlPath = saleService.executePythonScript();
         model.addAttribute("graphHtmlPath", graphHtmlPath);
-		return "SD/SD_purchaseData";
+		return "SD/SD_best";
 	}
 }
