+++ JM_ERP/src/main/resources/templates/navbar.html
@@ -56,7 +56,7 @@
                <li class="nav-item dropdown">
                     <a class="nav-link dropdown-toggle mainmenu-link navfontsize" data-bs-toggle="dropdown" href="#" role="button" aria-haspopup="true" aria-expanded="false">Marketing</a>
                     <div class="dropdown-menu">
-                        <a class="dropdown-item" href="http://localhost:8080/SD/member/list">영업 관리</a>
+                        <a class="dropdown-item" href="http://localhost:8080/SD/sale/best">영업 관리</a>
                         <a class="dropdown-item" href="http://localhost:8080/SD/member/list">고객 관리</a>
                         <a class="dropdown-item" href="http://localhost:8080/SD/seller/list">판매자 관리</a>
                         <a class="dropdown-item" href="http://localhost:8080/SD/product/nb">상품 관리</a>
